separate module registry tests
