+++ .github/workflows/test.yml
@@ -17,7 +17,7 @@ jobs:
         run: pip install -e ".[dev]"
 
       - name: Run tests with coverage
-        run: pytest tests/unit --cov=src/facts_experiment_builder --cov-report=xml
+        run: pytest tests/unit --ignore tests/unit/test_module_registry.py --cov=src/facts_experiment_builder --cov-report=xml
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4
